wip (save)

+++ src/main/java/request/RequestLoginAnonymous.java
@@ -18,6 +18,6 @@ public record RequestLoginAnonymous(StringChatFusion login) implements Request {
 
     @Override
     public ByteBuffer encode() {
-        return ByteBuffer.allocate(bufferLength()).put((byte) getOpCode().getOpCode()).put(login.encode());
+        return ByteBuffer.allocate(bufferLength()).put(getOpCode().getOpCode()).put(login.encode());
     }
 }
